+++ src/user/user.cpp
@@ -122,8 +122,8 @@ void termination_handler(int signal) {
 
 struct paxos_hdr {
     unsigned char op;
-    unsigned long slot;
-    unsigned long next;
+    unsigned long long slot;
+    unsigned long long next;
     int data_size;
 };
